requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
googlemaps==4.10.0
idna==3.10
kombu==5.4.2
mysqlclient==2.2.6
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
